+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Conversion tool: use repr when parsing numbers in excel files to ensure no precision is lost.
 ### Added
 - Audit: output transaction record in debug.
+- Bittrex: new data file format for deposits and withdrawals added.
 
 ## Version [0.1.4] Beta (2019-08-28)
 ### Fixed